Agregando mensajes de alertas cno Sweet Alert 2

+++ src/app/components/cart-app.component.ts
@@ -1,11 +1,11 @@
 import { Component, OnInit } from '@angular/core';
 import { ProductService } from '../services/product.service';
-import { Product } from '../models/product';
 import { CatalogComponent } from './catalog/catalog.component';
 import { CartItem } from '../models/cartItem';
 import { NavbarComponent } from "./navbar/navbar.component";
 import { Router, RouterOutlet } from '@angular/router';
 import { SharingDataService } from '../services/sharing-data.service';
+import Swal from 'sweetalert2';
 
 
 @Component({
@@ -16,7 +16,6 @@ import { SharingDataService } from '../services/sharing-data.service';
 })
 export class CartAppComponent implements OnInit{
 
-  products: Product[] = [];
   
   items: CartItem[] = [];
 
@@ -30,7 +29,6 @@ constructor(
 
 } 
   ngOnInit(): void {
-   this.products = this.service.findAll();
    this.items = JSON.parse(sessionStorage.getItem('cart') || '[]') ;
    this.calculateTotal();
    this.onDeleteCart();
@@ -60,23 +58,49 @@ constructor(
       this.calculateTotal();
       this.saveSession();
       this.router.navigate(['/cart'],{state:{items: this.items, total: this.total}});
+
+      Swal.fire({
+        title: "Shopping Cart",
+        text: "Nuevo producto agregado al carro",
+        icon: "success"
+      });
     })
   }
 
   onDeleteCart(): void {
     this.sharingDataService.idProductEventEmitter.subscribe(id => {
       console.log(id+' se ha ejecutado el evento idProductEventEmitter ')
-      this.items = this.items.filter( item => item.product.id !== id);
-      if(this.items.length == 0){
-        sessionStorage.removeItem('cart');
-        sessionStorage.clear();
-      }
-      this.calculateTotal();
-      this.saveSession();
-      this.router.navigateByUrl('/', {skipLocationChange: true}).then(()=>{
+      Swal.fire({
+        title: "Está seguro que desea eliminar?",
+        text: "Cuidado el ítem se eliminará del carro de compras!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Si, eliminar"
+      }).then((result) => {
+        if (result.isConfirmed) {
 
-        this.router.navigate(['/cart'],{state:{items: this.items, total: this.total}});
-      })
+
+
+          this.items = this.items.filter( item => item.product.id !== id);
+          if(this.items.length == 0){
+            sessionStorage.removeItem('cart');
+            sessionStorage.clear();
+          }
+          this.calculateTotal();
+          this.saveSession();
+          this.router.navigateByUrl('/', {skipLocationChange: true}).then(()=>{
+    
+            this.router.navigate(['/cart'],{state:{items: this.items, total: this.total}});
+          })
+          Swal.fire({
+            title: "Eliminado!",
+            text: "Se ha eliminado el ítem del carrito de compras.",
+            icon: "success"
+          });
+        }
+      });
     })
     }
